Split out BorderRenderHelper and add some notes to describe changes from upstream
